#wip renaming steps + post first event (still broken)
