+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step3.tsx
@@ -1,19 +1,62 @@
+import * as React from "react";
 import Typography from "@mui/material/Typography";
 import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 import { Box } from "@mui/system";
 import VC, { SolidVC, VCType } from '../../src/VC';
+import Image from "../../src/Image";
 import { useCallback, useState } from "react";
 
-import Image from "../../src/Image";
-import KadasterKnowledgeGraph from "../../src/KadasterKnowledgeGraph";
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableContainer from '@mui/material/TableContainer';
+import TableRow from '@mui/material/TableRow';
+import Paper from '@mui/material/Paper';
 import Link from "../../src/Link";
 
+function createData(
+  name: string,
+  value: string,
+) {
+  return { name, value };
+}
+
 export default function Step3({ stepNr = 3, handleNext, handleBack = () => { } }) {
-  const [loadedBRKVC, setLoadedBRKVC] = useState([] as any);
+  const [loadedBRPVC, setLoadedBRPVC] = useState({} as any);
+
+  const [rows, setRows] = useState([] as Array<any>);
 
   const updateVCs = useCallback(async (vcs: SolidVC[]) => {
-    setLoadedBRKVC(vcs);
+    // useCallback is important here. With each update from vcs
+    // <VC> will be re-rendered, and this will cause a new updateVCs
+    // resulting in an infinite loop.
+
+    if (vcs.length === 0) {
+      setLoadedBRPVC({});
+      return;
+    }
+
+    const vc = vcs[0];
+    setLoadedBRPVC(vc);
+
+    let certificateValidity;
+    // if type is string, it's a stringified JSON object
+    if (typeof vc.status === "string") {
+      certificateValidity = vc.status;
+    } else if (vc.status.verified) {
+      certificateValidity = "✅";
+    } else {
+      certificateValidity = "❌";
+    }
+
+    setRows([
+      createData('Naam', vc.vc.credentialSubject.naam),
+      createData('Geboortedatum', vc.vc.credentialSubject.geboorte.geboortedatum),
+      createData('Geboorteplaats', vc.vc.credentialSubject.geboorte.geboorteplaats),
+      createData('Geboorteland', vc.vc.credentialSubject.geboorte.geboorteland),
+      createData('Certificaat geldig?', certificateValidity),
+    ]);
   }, []);
 
   return (
@@ -22,38 +65,17 @@ export default function Step3({ stepNr = 3, handleNext, handleBack = () => { } }
         Start een nieuwe koopovereenkomst
       </Typography>
       <Typography variant="h2" color="text.primary" align="center">
-        {stepNr}. Koppel de woning die je wilt verkopen
+        {stepNr}. Koppel je persoonsgegevens aan deze koopovereenkomst
       </Typography>
 
-      <Typography variant="body1" color="text.primary" align="center">
-        Bij het Kadaster kan je jouw eigendomsbewijzen ophalen, zodat je vervolgens de woning kunt selecteren die je wilt verkopen.
-      </Typography>
-
-      <VC type={VCType.BRK} onChange={updateVCs} />
+      <VC type={VCType.BRP} onChange={updateVCs} />
 
       <hr />
-      {loadedBRKVC.length !== 0 ?
+      {Object.keys(loadedBRPVC).length === 0 ?
         <Box>
-          <Typography variant="body1" color="text.primary" align="center">
-            Je hebt je eigendomsgegevens opgeslagen in je datakluis. Kloppen de gegevens?
-          </Typography>
-          <Typography variant="body1" color="text.primary" align="center" fontWeight="bold" fontStyle="italic">
-            Certificaat geldig? {loadedBRKVC[0].status.verified ? "✅" : "❌"}
-          </Typography>
-          <Typography variant="body1" color="text.primary" align="center" sx={{
-            margin: "25px auto",
-            maxWidth: "600px",
-          }}>
-            Het volgende object ID van het perceel is gevonden in de verifiable credential van het Kadaster.
-            Maak gebruik van de Kadaster Knowledge Graph om informatie op te halen over je perceel.
-          </Typography>
-          <KadasterKnowledgeGraph objectId={(loadedBRKVC[0].vc.credentialSubject.eigendom.perceel.identificatie as number)} />
-        </Box>
-        :
-        <Box>
-          <Link href="/brkadaster" target="_blank">
+          <Link href="/brpersonen" target="_blank">
             <Image
-              src="/solid-quest/images/mijnkadaster.png"
+              src="/solid-quest/images/mijnoverheid.png"
               alt="Mijn Overheid Logo"
               width={400}
               height={180}
@@ -61,11 +83,40 @@ export default function Step3({ stepNr = 3, handleNext, handleBack = () => { } }
             />
           </Link>
         </Box>
+        :
+        <Box>
+          <Typography variant="body1" color="text.primary" align="center">
+            Je hebt je persoonsgegevens opgeslagen in je datakluis. Kloppen de gegevens?
+          </Typography>
+          {/* https://mui.com/material-ui/react-table/ */}
+          <TableContainer sx={{ marginY: '3rem' }} component={Paper}>
+            <Table sx={{ minWidth: 650 }} aria-label="simple table">
+              <TableBody>
+                {rows.map((row) => (
+                  <TableRow
+                    key={row.name}
+                    sx={{
+                      '&:last-child td, &:last-child th': {
+                        border: 0,
+                        fontStyle: 'italic',
+                        fontWeight: 'bold',
+                        height: '4rem',
+                      }
+                    }}
+                  >
+                    <TableCell>{row.name}</TableCell>
+                    <TableCell>{row.value}</TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+            </Table>
+          </TableContainer>
+        </Box>
       }
 
       <Stack direction="row" justifyContent="space-between">
         <Button variant="contained" onClick={handleBack}>Terug</Button>
-        {loadedBRKVC.length !== 0 && <Button variant="contained" onClick={handleNext}>Akkoord</Button>}
+        {Object.keys(loadedBRPVC).length !== 0 && <Button variant="contained" onClick={handleNext}>Akkoord</Button> }
       </Stack>
     </Box>
   );
