+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step7.tsx
@@ -1,24 +1,22 @@
-import Button from "@mui/material/Button";
-import Stack from "@mui/material/Stack";
+import * as React from "react";
 import Typography from "@mui/material/Typography";
+import Stack from "@mui/material/Stack";
+import Button from "@mui/material/Button";
 import { Box } from "@mui/system";
 
-export default function Step5({ stepNr = 5, handleNext, handleBack = () => { } }) {
+export default function Step7({ stepNr = 6, handleBack = () => {} }) {
   return (
     <Box sx={{ flex: 1 }}>
       <Typography variant="h1" color="text.primary" align="center">
         Start een nieuwe koopovereenkomst
       </Typography>
       <Typography variant="h2" color="text.primary" align="center">
-        {stepNr}. Koopdetails
+        {stepNr}. Tekenen koopovereenkomst
       </Typography>
 
       <Stack direction="row" justifyContent="space-between">
         <Button variant="contained" onClick={handleBack}>
           Terug
-        </Button>
-        <Button variant="contained" onClick={handleNext}>
-          Doorgaan
         </Button>
       </Stack>
     </Box>
