+++ koopovereenkomst-v3-simple/src/koek/KoekState.ts
@@ -32,6 +32,7 @@ export function initState(context?: object | undefined, value?: object | undefin
             ...internalState["@context"],
             ...context,
         };
+        console.log('update context', internalState);
     }
 
     return internalState;
